client: move client interface into types module
