+++ modules/client/src/index.ts
@@ -1,5 +1,4 @@
 import { connect, ConnextClient } from "./connext";
 import * as utils from "./lib/utils";
-import { ClientOptions } from "./types";
 
-export { ClientOptions, ConnextClient, connect, utils };
+export { ConnextClient, connect, utils };
