+++ modules/client/src/types.ts
@@ -4,8 +4,10 @@ import {
   AppState,
   ChannelProvider,
   ChannelState,
+  ClientOptions,
   GetConfigResponse,
   MultisigState,
+  Store,
 } from "@connext/types";
 import { Node as CFCoreTypes } from "@counterfactual/types";
 import { providers, utils, Wallet } from "ethers";
@@ -16,40 +18,8 @@ import { NodeApiClient } from "./node";
 export type BigNumber = utils.BigNumber;
 export const BigNumber = utils.BigNumber;
 
-export interface Store extends CFCoreTypes.IStoreService {
-  set(
-    pairs: {
-      path: string;
-      value: any;
-    }[],
-    shouldBackup?: Boolean,
-  ): Promise<void>;
-  restore(): Promise<{ path: string; value: any }[]>;
-}
-
-export interface ClientOptions {
-  // provider, passed through to CF node
-  ethProviderUrl: string;
-  // node information
-  nodeUrl: string; // ws:// or nats:// urls are supported
-
-  // signing options, include either a mnemonic directly
-  mnemonic?: string;
-
-  // or a channel provider
-  channelProvider?: ChannelProvider;
-  // function passed in by wallets to generate ephemeral keys
-  // used when signing applications
-  keyGen?: () => Promise<string>; // TODO: what will the type look like?
-  safeSignHook?: (state: ChannelState | AppState) => Promise<string>;
-  store: Store;
-  // TODO: state: string?
-  logLevel?: number; // see logger.ts for meaning, optional
-  // TODO: should be used in internal options? --> only if hardcoded
-  // nats communication config, client must provide
-  natsClusterId?: string;
-  natsToken?: string;
-}
+// TODO: define properly!!
+export interface ConnextStore {}
 
 export type InternalClientOptions = ClientOptions & {
   appRegistry: AppRegistry;
@@ -59,13 +29,10 @@ export type InternalClientOptions = ClientOptions & {
   ethProvider: providers.JsonRpcProvider;
   messaging: IMessagingService;
   multisigAddress: string;
-  network: utils.Network; // TODO: delete! use bos branch!
+  network: utils.Network;
   node: NodeApiClient;
   store: Store;
 };
-
-// TODO: define properly!!
-export interface ConnextStore {}
 
 ////////////////////////////////////////
 // NODE TYPES
