+++ modules/types/src/index.ts
@@ -2,6 +2,7 @@ export * from "./app";
 export * from "./basic";
 export * from "./channel";
 export * from "./channelProvider";
+export * from "./client";
 export * from "./convert";
 export * from "./inputs";
 export * from "./node";
